Utility file location refactoring (#134)
* Moving CLI menus to separate folder

* Moving files around so their location reflect where they are being used

Several utility files were being referenced from other parts of the application, indicating that they should be located at a higher level.
Also create a separate folder for DatabaseManager, since this is not directly related to the WorkloadSetupManager.

#patch
