+++ src/CLI/Menus/ConfigureEnvironment.cs
