+++ src/CLI/Menus/DeleteWorkloads.cs
