+++ src/CLI/Menus/EnableScaleUnitFeature.cs
