+++ src/CLI/Menus/EnableScaleUnitFeatureOnHub.cs
