+++ src/CLI/Menus/UpgradeWorkloadsDefinition.cs
