+++ src/CLI/Menus/WorkloadInstallationMenu.cs
