+++ src/CLI/Menus/WorkloadInstallationOptions/InstallWorkloads.cs
