+++ src/CLI/Menus/WorkloadInstallationOptions/WorkloadsInstallationStatus.cs
