+++ src/ScaleUnitManagement/ScaleUnitFeatureManager/Common/AddToolClientToSysAADClientTable.cs
@@ -1,5 +1,4 @@
 using System.Threading.Tasks;
-using ScaleUnitManagement.ScaleUnitFeatureManager.Utilities;
 using ScaleUnitManagement.Utilities;
 
 namespace ScaleUnitManagement.ScaleUnitFeatureManager.Common
