+++ src/ScaleUnitManagement/ScaleUnitFeatureManager/Common/UpdateMaxBatchSessions.cs
@@ -1,5 +1,5 @@
 using System.Threading.Tasks;
-using ScaleUnitManagement.ScaleUnitFeatureManager.Utilities;
+using ScaleUnitManagement.Utilities;
 
 namespace ScaleUnitManagement.ScaleUnitFeatureManager.Common
 {
