+++ src/ScaleUnitManagement/ScaleUnitFeatureManager/Utilities/CertificateStoreHelper.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Linq;
 using System.Security.Cryptography.X509Certificates;
+using ScaleUnitManagement.Utilities;
 
 namespace ScaleUnitManagement.ScaleUnitFeatureManager.Utilities
 {
