+++ src/ScaleUnitManagement/WorkloadSetupOrchestrator/Utilities/WorkloadExtensions.cs
